+++ Car Management/Controllers/VehicleController.cs
@@ -2,13 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Car_Management.Data;
-using Car_Management.Model;
-using Car_Management.Repository;
+using Cyberservice_management.Data;
+using Cyberservice_management.Model;
+using Cyberservice_management.Repository;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Car_Management.Controllers
+namespace Cyberservice_management.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
@@ -18,8 +19,8 @@ namespace Car_Management.Controllers
         private readonly ApplicationDbContext context;
 
         public VehicleController(
-         IVehicle _vehicle,
-        ApplicationDbContext _context)
+            IVehicle _vehicle,
+             ApplicationDbContext _context)
         {
             repository = _vehicle;
             context = _context;
@@ -27,6 +28,7 @@ namespace Car_Management.Controllers
 
         [HttpPost]
         [Route("AddVehicle")]
+        //[Authorize(Policy = "RequireAdministratorRole")]
         public IActionResult Post([FromBody] Vehicle vehicle)
         {
             if (!ModelState.IsValid)
@@ -49,37 +51,47 @@ namespace Car_Management.Controllers
             return Ok(vehicle);
         }
 
-        [HttpGet("{id}")]
-        public IActionResult Edit(int id)
+
+        [HttpGet]
+        [Route("GetVehicles")]
+        //[Authorize(Policy = "RequireLoggedIn")]
+        public Vehicle[] Get()
         {
-            Vehicle vehicle = repository.GetByid(id);
-            return Ok(vehicle);
+            return repository.GetVehicles().ToArray();
         }
 
-        [HttpPost]
+
+        [HttpPut]
         [Route("Edit")]
+       // [Authorize(Policy = "RequireAdministratorRole")]
         public IActionResult Edit([FromBody] Vehicle vehicle)
         {
             if (ModelState.IsValid)
             {
 
-                Vehicle EditedVehicle = new Vehicle();
-                repository.Update(EditedVehicle);
+                Vehicle EditVehicle = new Vehicle();
+                repository.Update(EditVehicle);
 
             }
             else
             {
                 return BadRequest();
             }
-            return Ok();
+            return Ok(new JsonResult("This Vehicle has been Updated Successfully"));
         }
 
-        [HttpDelete("{id}")]
-        public void Delete(int id)
+        [HttpDelete("Delete/{id}")]
+        //[Authorize(Policy = "RequireAdministratorRole")]
+        public IActionResult DeleteVehicle([FromRoute]int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             Vehicle vehicle = repository.GetByid(id);
             repository.Delete(vehicle);
-
+            return Ok(new JsonResult("Vehicle was Deleted Successfully"));
         }
     }
 }
